python/requirements.txt
openai>=1.75.0
python-dotenv>=1.0.1
websockets>=12.0
requests>=2.31.0
sounddevice>=0.4.6
